World Health Organization (WHO)

This dataset contains sample data for computing non-ordered summary measures of health inequality. It contains data from a household survey for two indicators, births attended by skilled health personnel (sba) and Diphtheria tetanus toxoid and pertussis (DTP3) immunization coverage, disaggregated by economic status. Both indicators are binary, (1) for those who had sba or dpt3 or (0) if the had not.

Usage

IndividualSample

Format

IndividualSample

A data frame with 17,848 rows and 10 columns:

id

individual identifier

psu

Primary Sample Unit (PSU)

strata

sampling strata

weight

sampling weight

subgroup

subgroup name

subgroup_order

subgroup order

sba

indicator estimate

dtp3

indicator estimate

favourable_indicator

favourable (1) or non-favourable (0) indicator

indicator_scale

scale of the indicator

Source

WHO Health Inequality Data Repositoryhttps://www.who.int/data/inequality-monitor/data

Details

Births attended by skilled health personnel is defined as a birth attended by skilled health personnel divided by the total number of live births to women aged 15-49 years occurring in the period prior to the survey. Skilled health personnel include doctors, nurses, midwives and other medically trained personnel, as defined according to each country. DPT3 is measured among one-year-olds and indicate those who have received three doses of the combined diphtheria, tetanus toxoid and pertussis containing vaccine in a given year.This is in line with the definition used by the Countdown to 2030 Collaboration, Demographic and Health Surveys (DHS), Multiple Indicator Cluster Surveys (MICS) and Reproductive Health Surveys (RHS).

Economic status is determined using a wealth index, which is based on owning selected assets and having access to certain services. The wealth index is divided into five equal subgroups (quintiles) that each account for 20% of the population. Economic status is an ordered dimension (meaning that the subgroups have an inherent ordering).

This dataset can be used to calculate ordered summary measures of health inequality, including: absolute concentration index (ACI), relative concentration index (RCI), slope index of inequality (SII) and relative index of inequality (RII).

Examples

head(IndividualSample)
#> # A tibble: 6 × 10
#>      id   psu strata weight subgroup             subgroup_order   sba  dtp3
#>   <dbl> <dbl>  <dbl>  <dbl> <chr>                         <dbl> <dbl> <dbl>
#> 1     1    88      1  0.352 Quintile 3                        3     1    NA
#> 2     2  1337     38  0.432 Quintile 1 (poorest)              1     1     1
#> 3     3   450     18  0.482 Quintile 2                        2     1    NA
#> 4     4  1692     56  0.407 Quintile 1 (poorest)              1     0    NA
#> 5     5   752     23  1.55  Quintile 5 (richest)              5     1    NA
#> 6     6  1033     30  2.43  Quintile 5 (richest)              5     1    NA
#> # ℹ 2 more variables: favourable_indicator <dbl>, indicator_scale <dbl>
summary(IndividualSample)
#>        id             psu           strata         weight       
#>  Min.   :    1   Min.   :   1   Min.   : 1.0   Min.   :0.06301  
#>  1st Qu.: 4463   1st Qu.: 474   1st Qu.:19.0   1st Qu.:0.35973  
#>  Median : 8924   Median :1066   Median :30.0   Median :0.84977  
#>  Mean   : 8924   Mean   :1016   Mean   :32.3   Mean   :0.95358  
#>  3rd Qu.:13386   3rd Qu.:1546   3rd Qu.:50.0   3rd Qu.:1.42668  
#>  Max.   :17848   Max.   :1970   Max.   :68.0   Max.   :5.40138  
#>                                                                 
#>    subgroup         subgroup_order       sba              dtp3      
#>  Length:17848       Min.   :1.000   Min.   :0.0000   Min.   :0.000  
#>  Class :character   1st Qu.:1.000   1st Qu.:1.0000   1st Qu.:1.000  
#>  Mode  :character   Median :3.000   Median :1.0000   Median :1.000  
#>                     Mean   :2.761   Mean   :0.7667   Mean   :0.751  
#>                     3rd Qu.:4.000   3rd Qu.:1.0000   3rd Qu.:1.000  
#>                     Max.   :5.000   Max.   :1.0000   Max.   :1.000  
#>                                                      NA's   :14313  
#>  favourable_indicator indicator_scale
#>  Min.   :1            Min.   :100    
#>  1st Qu.:1            1st Qu.:100    
#>  Median :1            Median :100    
#>  Mean   :1            Mean   :100    
#>  3rd Qu.:1            3rd Qu.:100    
#>  Max.   :1            Max.   :100    
#>