Absolute concentration index (ACI)

The absolute concentration index (ACI) is an absolute measure of inequality that indicates the extent to which an indicator is concentrated among disadvantaged or advantaged subgroups, on an absolute scale.

Usage

aci(
  est,
  subgroup_order,
  pop = NULL,
  weight = NULL,
  psu = NULL,
  strata = NULL,
  fpc = NULL,
  lmin = NULL,
  lmax = NULL,
  conf.level = 0.95,
  force = FALSE,
  ...
)

Arguments

est

The indicator estimate. Estimates must be available for all subgroups/individuals (unless force=TRUE).

subgroup_order

The order of subgroups/individuals in an increasing sequence.

pop

For disaggregated data, the number of people within each subgroup. This must be available for all subgroups.

weight

The individual sampling weight, for individual-level data from a survey. This must be available for all individuals.

psu

Primary sampling unit, for individual-level data from a survey.

strata

Strata, for individual-level data from a survey.

fpc

Finite population correction, for individual-level data from a survey where sample size is large relative to population size.

lmin

Minimum limit for bounded indicators (i.e., variables that have a finite upper and/or lower limit).

lmax

Maximum limit for bounded indicators (i.e., variables that have a finite upper and/or lower limit).

conf.level

Confidence level of the interval. Default is 0.95 (95%).

force

TRUE/FALSE statement to force calculation with missing indicator estimate values.

...

Further arguments passed to or from other methods.

Value

The estimated ACI value, corresponding estimated standard error, and confidence interval as a data.frame.

Details

ACI can be calculated using disaggregated data and individual-level data. Subgroups in disaggregated data are weighted according to their population share, while individuals are weighted by sample weight in the case of data from surveys.

The calculation of ACI is based on a ranking of the whole population from the most disadvantaged subgroup (at rank 0) to the most advantaged subgroup (at rank 1), which is inferred from the ranking and size of the subgroups. ACI can be calculated as twice the covariance between the health indicator and the relative rank. Given the relationship between covariance and ordinary least squares regression, ACI can be obtained from a regression of a transformation of the health variable of interest on the relative rank. For more information on this inequality measure see Schlotheuber (2022) below.

Interpretation: ACI is 0 if there is no inequality. The larger the absolute value of ACI, the higher the level of inequality. Positive values indicate a concentration of the indicator among advantaged subgroups, and negative values indicate a concentration of the indicator among disadvantaged subgroups.

Type of summary measure: Complex; absolute; weighted

Applicability: Ordered dimension of inequality with more than two subgroups

Warning: The confidence intervals are approximate and might be biased.

References

Schlotheuber, A, Hosseinpoor, AR. Summary measures of health inequality: A review of existing measures and their application. Int J Environ Res Public Health. 2022;19(6):3697. doi:10.3390/ijerph19063697.

Examples

# example code
data(IndividualSample)
head(IndividualSample)
#>   id  psu strata   weight             subgroup subgroup_order sba dtp3
#> 1  1   88      1 0.351672           Quintile 3              3   1   NA
#> 2  2 1337     38 0.431545 Quintile 1 (poorest)              1   1    1
#> 3  3  450     18 0.482483           Quintile 2              2   1   NA
#> 4  4 1692     56 0.407390 Quintile 1 (poorest)              1   0   NA
#> 5  5  752     23 1.547062 Quintile 5 (richest)              5   1   NA
#> 6  6 1033     30 2.429523 Quintile 5 (richest)              5   1   NA
#>   favourable_indicator indicator_scale
#> 1                    1             100
#> 2                    1             100
#> 3                    1             100
#> 4                    1             100
#> 5                    1             100
#> 6                    1             100
with(IndividualSample,
     aci(est = sba,
         subgroup_order = subgroup_order,
         weight = weight,
         psu = psu,
         strata = strata))
#>   measure   estimate          se    lowerci    upperci
#> 1     aci 0.07801353 0.002950566 0.07223053 0.08379654
# example code
data(OrderedSample)
head(OrderedSample)
#>                                         indicator
#> 1 Births attended by skilled health personnel (%)
#> 2 Births attended by skilled health personnel (%)
#> 3 Births attended by skilled health personnel (%)
#> 4 Births attended by skilled health personnel (%)
#> 5 Births attended by skilled health personnel (%)
#>                           dimension             subgroup subgroup_order
#> 1 Economic status (wealth quintile) Quintile 1 (poorest)              1
#> 2 Economic status (wealth quintile)           Quintile 2              2
#> 3 Economic status (wealth quintile)           Quintile 3              3
#> 4 Economic status (wealth quintile)           Quintile 4              4
#> 5 Economic status (wealth quintile) Quintile 5 (richest)              5
#>   estimate        se population setting_average favourable_indicator
#> 1 75.60530 1.5996131   2072.436        91.59669                    1
#> 2 91.01997 1.1351504   2112.204        91.59669                    1
#> 3 96.03959 0.6461946   1983.059        91.59669                    1
#> 4 97.04223 0.5676206   2052.124        91.59669                    1
#> 5 99.22405 0.2237683   1884.510        91.59669                    1
#>   ordered_dimension indicator_scale
#> 1                 1             100
#> 2                 1             100
#> 3                 1             100
#> 4                 1             100
#> 5                 1             100
with(OrderedSample,
     aci(est = estimate,
         subgroup_order = subgroup_order,
         pop = population))
#>   measure estimate       se  lowerci  upperci
#> 1     aci  4.30052 1.284031 1.783865 6.817174