Between-group standard deviation (BGSD)

Between-group standard deviation (BGSD) is an absolute measure of inequality that considers all population subgroups. Subgroups are weighted according to their population share.

Usage

bgsd(
  est,
  se = NULL,
  pop,
  scaleval = NULL,
  sim = NULL,
  seed = 123456,
  force = FALSE,
  ...
)

Arguments

est

The subgroup estimate. Estimates must be available for at least 85% of subgroups.

se

The standard error of the subgroup estimate. If this is missing, 95% confidence intervals cannot be calculated.

pop

The number of people within each subgroup.Population size must be available for all subgroups.

scaleval

The scale of the indicator. For example, the scale of an indicator measured as a percentage is 100. The scale of an indicator measured as a rate per 1000 population is 1000. If this is missing, 95% confidence intervals cannot be calculated.

sim

The number of simulations to estimate confidence intervals. Default is 100.

seed

The random number generator (RNG) state for the confidence interval simulation. Default is 123456.

force

TRUE/FALSE statement to force calculation when more than 85% of subgroup estimates are missing.

...

Further arguments passed to or from other methods.

Value

The estimated BGSD value, corresponding estimated standard error, and confidence interval as a data.frame.

Details

BGSD is calculated as the square root of the weighted average of squared differences between the subgroup estimates and the setting average. Squared differences are weighted by each subgroup’s population share. For more information on this inequality measure see Schlotheuber (2022) below.

95% confidence intervals are calculated using a Monte Carlo simulation-based method. The dataset is simulated a large number of times (e.g. 100), with the mean and standard error of each simulated dataset being the same as the original dataset. BGSD is calculated for each of the simulated sample datasets. The 95% confidence intervals are based on the 2.5th and 97.5th percentiles of the BGSD results. See Ahn (2019) below for further information.

Interpretation: BGSD has only positive values, with larger values indicating higher levels of inequality. BGSD is 0 if there is no inequality. BGSD has the same unit as the health indicator.

Type of summary measure: Complex; absolute; weighted

Applicability: Non-ordered dimensions of inequality with more than two subgroups

References

Schlotheuber, A, Hosseinpoor, AR. Summary measures of health inequality: A review of existing measures and their application. Int J Environ Res Public Health. 2022;19(6):3697. doi:10.3390/ijerph19063697.

Ahn J, Harper S, Yu M, Feuer EJ, Liu B. Improved Monte Carlo methods for estimating confidence intervals for eleven commonly used health disparity measures. PLoS One. 2019 Jul 1;14(7).

Examples

# example code
data(NonorderedSample)
head(NonorderedSample)
#>                                         indicator          dimension
#> 1 Births attended by skilled health personnel (%) Subnational region
#> 2 Births attended by skilled health personnel (%) Subnational region
#> 3 Births attended by skilled health personnel (%) Subnational region
#> 4 Births attended by skilled health personnel (%) Subnational region
#> 5 Births attended by skilled health personnel (%) Subnational region
#> 6 Births attended by skilled health personnel (%) Subnational region
#>          subgroup  estimate        se population setting_average
#> 1            aceh  95.11784 1.5384434  230.20508        91.59669
#> 2            bali 100.00000 0.0000000  149.46272        91.59669
#> 3 bangka balitung  97.41001 1.2676437   55.66533        91.59669
#> 4          banten  80.35694 3.5440531  451.26550        91.59669
#> 5        bengkulu  94.25756 2.7740061   70.17540        91.59669
#> 6    central java  98.56168 0.6476116 1221.94446        91.59669
#>   favourable_indicator ordered_dimension indicator_scale reference_subgroup
#> 1                    1                 0             100                  1
#> 2                    1                 0             100                  0
#> 3                    1                 0             100                  0
#> 4                    1                 0             100                  0
#> 5                    1                 0             100                  0
#> 6                    1                 0             100                  0
with(NonorderedSample,
     bgsd(est = estimate,
          se = se,
          pop = population,
          scaleval = indicator_scale))
#>   measure estimate se  lowerci  upperci
#> 1    bgsd  7.10072 NA 6.327117 8.540935