Difference (D)

The difference (D) is an absolute measure of inequality that shows the difference in an indicator between two population subgroups. For more information on this inequality measure see Schlotheuber (2022) below.

Usage

d(
  est,
  se = NULL,
  favourable_indicator,
  ordered_dimension,
  subgroup_order = NULL,
  reference_subgroup = NULL,
  conf.level = 0.95,
  ...
)

Arguments

est

The subgroup estimate. Estimates must be available for the two subgroups being compared.

se

The standard error of the subgroup estimate. If this is missing, confidence intervals of D cannot be calculated.

favourable_indicator

Records whether the indicator is favourable (1) or adverse (0). Favourable indicators measure desirable health events where the ultimate goal is to achieve a maximum level (such as skilled birth attendance). Adverse indicators measure undesirable health events where the ultimate goal is to achieve a minimum level (such as under-five mortality rate).

ordered_dimension

Records whether the dimension is ordered (1) or non-ordered (0). Ordered dimensions have subgroup with a natural order (such as economic status). Non-ordered or binary dimensions do not have a natural order (such as subnational region or sex).

subgroup_order

The order of subgroups in an increasing sequence. Required if the dimension is ordered (ordered_dimension=1).

reference_subgroup

Identifies a reference subgroup with the value of 1, if the dimension is non-ordered or binary.

conf.level

Confidence level of the interval. Default is 0.95 (95%).

...

Further arguments passed to or from other methods.

Value

The estimated D value, corresponding estimated standard error, and confidence interval as a data.frame.

Details

D is calculated as D = y1 - y2, where y1 and y2 indicate the estimates for subgroups 1 and 2. The selection of the two subgroups depends on the characteristics of the inequality dimension and the purpose of the analysis. In addition, the direction of the calculation may depend on the indicator type (favourable or adverse). Please see specifications of how y1 and y2 are identified below.

Ordered dimension: Favourable indicator: Most-advantaged subgroup - Least-advantaged subgroup Adverse indicator: Least-advantaged subgroup - Most-advantaged subgroup

Non-ordered dimension: No reference group & favourable indicator: Highest estimate - Lowest estimate No reference group & adverse indicator: Lowest estimate - Highest estimate Reference group & favourable indicator: Reference estimate - Lowest estimate Reference group & adverse indicator: Lowest estimate - Reference estimate

Interpretation: Greater absolute values indicate higher levels of inequality. D is 0 if there is no inequality.

Type of summary measure: Simple; relative; unweighted

Applicability: Any dimension of inequality

Warning: The confidence intervals are approximate and might be biased. See Ahn et al. (2018) below for further information about the standard error formula.

References

Schlotheuber, A, Hosseinpoor, AR. Summary measures of health inequality: A review of existing measures and their application. Int J Environ Res Public Health. 2022;19(6):3697. doi:10.3390/ijerph19063697.

Ahn J, Harper S, Yu M, Feuer EJ, Liu B, Luta G. Variance estimation and confidence intervals for 11 commonly used health disparity measures. JCO Clin Cancer Inform. 2018;2:1-19. doi:10.1200/CCI.18.00031.

Examples

# example code
data(NonorderedSample)
head(NonorderedSample)
#>                                         indicator          dimension
#> 1 Births attended by skilled health personnel (%) Subnational region
#> 2 Births attended by skilled health personnel (%) Subnational region
#> 3 Births attended by skilled health personnel (%) Subnational region
#> 4 Births attended by skilled health personnel (%) Subnational region
#> 5 Births attended by skilled health personnel (%) Subnational region
#> 6 Births attended by skilled health personnel (%) Subnational region
#>          subgroup  estimate        se population setting_average
#> 1            aceh  95.11784 1.5384434  230.20508        91.59669
#> 2            bali 100.00000 0.0000000  149.46272        91.59669
#> 3 bangka balitung  97.41001 1.2676437   55.66533        91.59669
#> 4          banten  80.35694 3.5440531  451.26550        91.59669
#> 5        bengkulu  94.25756 2.7740061   70.17540        91.59669
#> 6    central java  98.56168 0.6476116 1221.94446        91.59669
#>   favourable_indicator ordered_dimension indicator_scale reference_subgroup
#> 1                    1                 0             100                  1
#> 2                    1                 0             100                  0
#> 3                    1                 0             100                  0
#> 4                    1                 0             100                  0
#> 5                    1                 0             100                  0
#> 6                    1                 0             100                  0
with(NonorderedSample,
     d(est = estimate,
       se = se,
       favourable_indicator = favourable_indicator,
       ordered_dimension = ordered_dimension,
       reference_subgroup = reference_subgroup))
#>   measure estimate       se lowerci  upperci
#> 1       d 30.92258 7.217724 16.7761 45.06906