Mean difference from the best-performing subgroup (unweighted) (MDBU)

The Mean Difference from the Best-Performing Subgroup (MDB) is an absolute measure of inequality that shows the mean difference between each population subgroup and the subgroup with the best estimate. For the unweighted version (MDBU), all subgroups are weighted equally.

Usage

mdbu(
  est,
  se = NULL,
  favourable_indicator,
  scaleval,
  sim = NULL,
  seed = 123456,
  ...
)

Arguments

est

The subgroup estimate. Estimates must be available for at least 85% of subgroups.

se

The standard error of the subgroup estimate. If this is missing, 95% confidence intervals of MDBU cannot be calculated.

favourable_indicator

Records whether the indicator is favourable (1) or non-favourable (0). Favourable indicators measure desirable health events where the ultimate goal is to achieve a maximum level (such as skilled birth attendance). Non-favourable indicators measure undesirable health events where the ultimate goal is to achieve a minimum level (such as under-five mortality rate).

scaleval

The scale of the indicator. For example, the scale of an indicator measured as a percentage is 100. The scale of an indicator measured as a rate per 1000 population is 1000.

sim

The number of simulations to estimate 95% confidence intervals. Default is 100.

seed

The random number generator (RNG) state for the 95% confidence interval simulation. Default is 123456.

...

Further arguments passed to or from other methods.

Value

The estimated MDBU value, corresponding estimated standard error, and confidence interval as a data.frame.

Details

The unweighted version (MDBU) is calculated as the average of absolute differences between the subgroup estimates and the estimate for the best-performing subgroup, divided by the number of subgroups. For more information on this inequality measure see Schlotheuber, A., & Hosseinpoor, A. R. (2022) below.

95% confidence intervals are calculated using a methodology of simulated estimates. The dataset is simulated a large number of times (e.g., 100) and MDBU is calculated for each of the simulated samples. The 95% confidence intervals are based on the 2.5th and 97.5th percentiles of the MDBU results.

Interpretation: MDBU only has positive values, with larger values indicating higher levels of inequality. MDBU is zero if there is no inequality.

Type of summary measure: Complex; absolute; non-weighted

Applicability: Non-ordered; more than two subgroups

References

Schlotheuber, A., & Hosseinpoor, A. R. (2022). Summary measures of health inequality: A review of existing measures and their application. International Journal of Environmental Research and Public Health, 19 (6), 3697.

Examples

# example code
data(NonorderedSample)
head(NonorderedSample)
#>                                         indicator          dimension
#> 1 Births attended by skilled health personnel (%) Subnational region
#> 2 Births attended by skilled health personnel (%) Subnational region
#> 3 Births attended by skilled health personnel (%) Subnational region
#> 4 Births attended by skilled health personnel (%) Subnational region
#> 5 Births attended by skilled health personnel (%) Subnational region
#> 6 Births attended by skilled health personnel (%) Subnational region
#>          subgroup  estimate        se population setting_average
#> 1            aceh  95.11784 1.5384434  230.20508        91.59669
#> 2            bali 100.00000 0.0000000  149.46272        91.59669
#> 3 bangka balitung  97.41001 1.2676437   55.66533        91.59669
#> 4          banten  80.35694 3.5440531  451.26550        91.59669
#> 5        bengkulu  94.25756 2.7740061   70.17540        91.59669
#> 6    central java  98.56168 0.6476116 1221.94446        91.59669
#>   favourable_indicator ordered_dimension indicator_scale reference_subgroup
#> 1                    1                 0             100                  1
#> 2                    1                 0             100                  0
#> 3                    1                 0             100                  0
#> 4                    1                 0             100                  0
#> 5                    1                 0             100                  0
#> 6                    1                 0             100                  0
with(NonorderedSample,
     mdbu(est = estimate,
          se,
          favourable_indicator,
          scaleval = indicator_scale
         )
     )
#>   measure estimate  lowerci  upperci
#> 1    mdbu  10.3264 9.253222 11.10566