Mean difference from a reference subgroup (weighted) (MDRW)

An absolute measure of inequality that shows the mean difference between each population subgroup and a reference subgroup. In the weighted version (MDRW), subgroups are weighted according to their population share.

Usage

mdrw(
  pop,
  est,
  se = NULL,
  scaleval,
  reference_subgroup,
  sim = NULL,
  seed = 123456,
  ...
)

Arguments

pop

The number of people within each subgroup. Population size must be available for all subgroups.

est

The subgroup estimate. Estimates must be available for all subgroups.

se

The standard error of the subgroup estimate. If this is missing, 95% confidence intervals of MDRW cannot be calculated.

scaleval

The scale of the indicator. For example, the scale of an indicator measured as a percentage is 100. The scale of an indicator measured as a rate per 1000 population is 1000.

reference_subgroup

Identifies a reference subgroup with the value of 1.

sim

The number of simulations to estimate 95% confidence intervals

seed

The random number generator (RNG) state for the 95% confidence interval simulation

...

Further arguments passed to or from other methods.

Value

The estimated MDRW value, corresponding estimated standard error, and confidence interval as a data.frame.

Details

The weighted version (MDRW) is calculated as the weighted average of absolute differences between the subgroup estimates and the estimate for the reference subgroup. Absolute differences are weighted by each subgroup’s population share. For more information on this inequality measure see Schlotheuber, A., & Hosseinpoor, A. R. (2022) below.

95% confidence intervals are calculated using a methodology of simulated estimates. The dataset is simulated a large number of times (e.g., 100) and MDRW is calculated for each of the simulated samples. The 95% confidence intervals are based on the 2.5th and 97.5th percentiles of the MDRW results.

Interpretation: MDRW only has positive values, with larger values indicating higher levels of inequality. MDRW is zero if there is no inequality.

Type of summary measure: Complex; absolute; weighted

Applicability: Non-ordered; more than two subgroups

References

Schlotheuber, A., & Hosseinpoor, A. R. (2022). Summary measures of health inequality: A review of existing measures and their application. International journal of environmental research and public health, 19 (6), 3697.

Examples

# example code
data(NonorderedSample)
head(NonorderedSample)
#> # A tibble: 6 × 11
#>   indicator         dimension subgroup estimate    se population setting_average
#>   <chr>             <chr>     <chr>       <dbl> <dbl>      <dbl>           <dbl>
#> 1 Births attended … Subnatio… aceh         95.1 1.54       230.             91.6
#> 2 Births attended … Subnatio… bali        100   0          149.             91.6
#> 3 Births attended … Subnatio… bangka …     97.4 1.27        55.7            91.6
#> 4 Births attended … Subnatio… banten       80.4 3.54       451.             91.6
#> 5 Births attended … Subnatio… bengkulu     94.3 2.77        70.2            91.6
#> 6 Births attended … Subnatio… central…     98.6 0.648     1222.             91.6
#> # ℹ 4 more variables: favourable_indicator <dbl>, ordered_dimension <dbl>,
#> #   indicator_scale <dbl>, reference_subgroup <dbl>
with(NonorderedSample,
     mdrw(pop = population,
          est = estimate,
          se = se,
          scaleval = indicator_scale,
          reference_subgroup
         )
     )
#>      measure estimate lowerci  upperci
#> 2.5%    mdrw  5.64322 4.66525 7.056331