Population attributable fraction (PAF)

The population attributable fraction (PAF) is a relative measure of inequality that shows the potential improvement in the average of an indicator, in relative terms, that could be achieved if all population subgroups had the same level of the indicator as a reference group.

Usage

paf(
  pop = NULL,
  est,
  ordered_dimension,
  subgroup_order = NULL,
  setting_average = NULL,
  favourable_indicator,
  scaleval,
  conf.level = 0.95,
  ...
)

Arguments

pop

The number of people within each subgroup. Population size must be available for all subgroups.

est

The subgroup estimate. Estimates must be available for all subgroups.

ordered_dimension

Records whether the dimension is ordered (1) or not (0).

subgroup_order

The order of subgroups in an increasing sequence, if the dimension is ordered (ordered dimension=1).

setting_average

The reported setting average. Setting average must be unique for each setting, year, indicator combination. If population is not specified for all subgroups, the setting average is used.

favourable_indicator

Records whether the indicator is favourable (1) or non-favourable (0). Favourable indicators measure desirable health events where the ultimate goal is to achieve a maximum level (such as skilled birth attendance). Non-favourable indicators measure undesirable health events where the ultimate goal is to achieve a minimum level (such as under-five mortality rate).

scaleval

The scale of the indicator. For example, the scale of an indicator measured as a percentage is 100. The scale of an indicator measured as a rate per 1000 population is 1000.

conf.level

Confidence level of the interval. Default is 0.95 (95%).

...

Further arguments passed to or from other methods.

Value

The estimated PAF value, corresponding estimated standard error, and confidence interval as a data.frame.

Details

PAF is calculated as the difference between the estimate for the reference subgroup and the setting average, divided by the setting average and multiplied by 100. For more information on this inequality measure see Schlotheuber, A., & Hosseinpoor, A. R. (2022) below.

If the indicator is favourable and PAF < 0, then PAF is replaced with 0. If the indicator is adverse and PAF > 0, then PAF is replaced with 0. The selection of the reference subgroup depends on the characteristics of the inequality dimension and the indicator type. It is the most-advantaged subgroup for ordered dimensions. For non-ordered dimensions, it is the subgroup with the highest estimate for favourable indicators and is the subgroup with the lowest estimate for adverse indicators.

Interpretation: PAF assumes positive values for favourable indicators and negative values for non-favourable (adverse) indicators. The larger the absolute value of PAF, the higher the level of inequality. PAF is zero if no further improvement can be achieved (i.e., if all subgroups have reached the same level of the indicator as the reference subgroup or surpassed that level).

Type of summary measure: Complex; relative; weighted

Applicability: Any

Warning: The confidence intervals are approximate and might be biased. See Walter S.D. (1978) below for further information on the standard error formula.

References

Schlotheuber, A., & Hosseinpoor, A. R. (2022). Summary measures of health inequality: A review of existing measures and their application. International Journal of Environmental Research and Public Health, 19 (6), 3697.

Walter, S. D. (1978). “Calculation of Attributable Risks from Epidemiological Data.” International Journal of Epidemiology 7 (2): 175–82.

Examples

# example code
data(OrderedSample)
head(OrderedSample)
#>                                         indicator
#> 1 Births attended by skilled health personnel (%)
#> 2 Births attended by skilled health personnel (%)
#> 3 Births attended by skilled health personnel (%)
#> 4 Births attended by skilled health personnel (%)
#> 5 Births attended by skilled health personnel (%)
#>                           dimension             subgroup subgroup_order
#> 1 Economic status (wealth quintile) Quintile 1 (poorest)              1
#> 2 Economic status (wealth quintile)           Quintile 2              2
#> 3 Economic status (wealth quintile)           Quintile 3              3
#> 4 Economic status (wealth quintile)           Quintile 4              4
#> 5 Economic status (wealth quintile) Quintile 5 (richest)              5
#>   estimate        se population setting_average favourable_indicator
#> 1 75.60530 1.5996131   2072.436        91.59669                    1
#> 2 91.01997 1.1351504   2112.204        91.59669                    1
#> 3 96.03959 0.6461946   1983.059        91.59669                    1
#> 4 97.04223 0.5676206   2052.124        91.59669                    1
#> 5 99.22405 0.2237683   1884.510        91.59669                    1
#>   ordered_dimension indicator_scale
#> 1                 1             100
#> 2                 1             100
#> 3                 1             100
#> 4                 1             100
#> 5                 1             100
with(OrderedSample,
     paf(pop = population,
         est = estimate,
         ordered_dimension = ordered_dimension,
         subgroup_order = subgroup_order,
         favourable_indicator = favourable_indicator,
         scaleval = indicator_scale
         )
     )
#>   measure estimate          se  lowerci  upperci
#> 1     paf 8.327116 0.003681919 8.319899 8.334332