Slope index of inequality (SII)

The slope index of inequality (SII) is an absolute measure of inequality that represents the difference in estimated indicator values between the most-advantaged and most-disadvantaged, while taking into consideration the situation in all other subgroups/individuals – using an appropriate regression model. SII can be calculated using both disaggregated data and individual-level data. Subgroups in disaggregated data are weighted according to their population share, while individuals are weighted by sample weight in the case of data from surveys.

sii(
  est,
  subgroup_order,
  pop = NULL,
  scaleval = NULL,
  weight = NULL,
  psu = NULL,
  strata = NULL,
  fpc = NULL,
  conf.level = 0.95,
  linear = FALSE,
  force = FALSE,
  ...
)

Arguments

est

The subgroup estimate. Estimates must be available for all subgroups.

subgroup_order

The order of subgroups in an increasing sequence.

pop

The number of people within each subgroup. Population size must be available for all subgroups.

scaleval

The scale of the indicator. For example, the scale of an indicator measured as a percentage is 100. The scale of an indicator measured as a rate per 1000 population is 1000.

weight

Individual sampling weight (required if data come from a survey)

psu

Primary sampling unit (required if data come from a survey)

strata

Strata (required if data come from a survey)

fpc

Finite population correction

conf.level

confidence level of the interval.

linear

TRUE/FALSE statement to specify the use of a linear regression model for SII estimation (default is logistic regression)

force

TRUE/FALSE statement to force calculation with missing indicator estimate values.

...

Further arguments passed to or from other methods.

Value

The estimated SII value, corresponding estimated standard error, and confidence interval as a data.frame.

Details

To calculate SII, a weighted sample of the whole population is ranked from the most-disadvantaged subgroup (at rank 0) to the most-advantaged subgroup (at rank 1). This ranking is weighted, accounting for the proportional distribution of the population within each subgroup. The indicator of interest is then regressed against this relative rank using an appropriate regression model (e.g., a generalized linear model with logit link), and the predicted values of the indicator are calculated for the two extremes (rank 1 and rank 0). The difference between the predicted values at rank 1 and rank 0 (covering the entire distribution) generates the SII value. For more information on this inequality measure see Schlotheuber, A., & Hosseinpoor, A. R. (2022) below.

Interpretation: SII is zero if there is no inequality. Greater absolute values indicate higher levels of inequality. For favourable indicators, positive values indicate a concentration of the indicator among the advantaged, while negative values indicate a concentration of the indicator among the disadvantaged. For adverse indicators, it is the reverse: positive values indicate a concentration of the indicator among the disadvantaged, while negative values indicate a concentration of the indicator among the advantaged.

Type of summary measure: Complex; absolute; weighted

Applicability: Ordered; more than two subgroups

Warning: The confidence intervals are approximate and might be biased.

References

Schlotheuber, A., & Hosseinpoor, A. R. (2022). Summary measures of health inequality: A review of existing measures and their application. International journal of environmental research and public health, 19 (6), 3697.

Examples

# example code
data(IndividualSample)
head(IndividualSample)
#> # A tibble: 6 × 10
#>      id   psu strata weight subgroup             subgroup_order   sba  dtp3
#>   <dbl> <dbl>  <dbl>  <dbl> <chr>                         <dbl> <dbl> <dbl>
#> 1     1    88      1  0.352 Quintile 3                        3     1    NA
#> 2     2  1337     38  0.432 Quintile 1 (poorest)              1     1     1
#> 3     3   450     18  0.482 Quintile 2                        2     1    NA
#> 4     4  1692     56  0.407 Quintile 1 (poorest)              1     0    NA
#> 5     5   752     23  1.55  Quintile 5 (richest)              5     1    NA
#> 6     6  1033     30  2.43  Quintile 5 (richest)              5     1    NA
#> # ℹ 2 more variables: favourable_indicator <dbl>, indicator_scale <dbl>
with(IndividualSample,
     sii(est = sba,
         subgroup_order = subgroup_order,
         weight = weight,
         psu = psu,
         strata = strata
         )
     )
#>   measure  estimate         se   lowerci   upperci
#> 1     sii 0.4941437 0.01753301 0.4597797 0.5285078