World Health Organization (WHO)

This dataset contains sample data for computing non-ordered summary measures of health inequality. It contains data from a household survey for the proportion of births attended by skilled health personnel disaggregated by subnational region.

NonorderedSample

Format

NonorderedSample

A data frame with 34 rows and 11 columns:

indicator

indicator name

dimension

dimension of inequality

subgroup

population subgroup within a given dimension of inequality

estimate

subgroup estimate

se

standard error of the subgroup estimate

population

number of people within each subgroup

setting_average

indicator average for the setting

favourable_indicator

favourable (1) or non-favourable (0) indicator

ordered_dimension

ordered (1) or non-ordered (0) dimension

indicator_scale

scale of the indicator

reference_subgroup

reference subgroup

Source

WHO Health Inequality Data Repositoryhttps://www.who.int/data/inequality-monitor/data

Details

The proportion of births attended by skilled health personnel is calculated as the number of births attended by skilled health personnel divided by the total number of live births to women aged 15-49 years occurring in the period prior to the survey.

Skilled health personnel include doctors, nurses, midwives and other medically trained personnel, as defined according to each country. This is in line with the definition used by the Countdown to 2030 Collaboration, Demographic and Health Surveys (DHS), Multiple Indicator Cluster Surveys (MICS) and Reproductive Health Surveys (RHS).

Subnational regions are defined using country-specific criteria. Subnational region is a non-ordered dimension (meaning that the subgroups do not have an inherent ordering).

This dataset can be used to calculate non-ordered summary measures of health inequality, including: between-group variance (BGV), between-group standard deviation (BGSD), coefficient of variation (COV), mean difference from mean (MDM), index of disparity (IDIS), Theil index (TI) and mean log deviation (MLD). It can also be used to calculate the impact measures population attributable risk (PAR) and population attributable fraction (PAF).

Examples

head(NonorderedSample)
#> # A tibble: 6 × 11
#>   indicator         dimension subgroup estimate    se population setting_average
#>   <chr>             <chr>     <chr>       <dbl> <dbl>      <dbl>           <dbl>
#> 1 Births attended … Subnatio… aceh         95.1 1.54       230.             91.6
#> 2 Births attended … Subnatio… bali        100   0          149.             91.6
#> 3 Births attended … Subnatio… bangka …     97.4 1.27        55.7            91.6
#> 4 Births attended … Subnatio… banten       80.4 3.54       451.             91.6
#> 5 Births attended … Subnatio… bengkulu     94.3 2.77        70.2            91.6
#> 6 Births attended … Subnatio… central…     98.6 0.648     1222.             91.6
#> # ℹ 4 more variables: favourable_indicator <dbl>, ordered_dimension <dbl>,
#> #   indicator_scale <dbl>, reference_subgroup <dbl>
summary(NonorderedSample)
#>   indicator          dimension           subgroup            estimate     
#>  Length:34          Length:34          Length:34          Min.   : 64.20  
#>  Class :character   Class :character   Class :character   1st Qu.: 86.76  
#>  Mode  :character   Mode  :character   Mode  :character   Median : 91.20  
#>                                                           Mean   : 89.67  
#>                                                           3rd Qu.: 96.56  
#>                                                           Max.   :100.00  
#>        se          population      setting_average favourable_indicator
#>  Min.   :0.000   Min.   :  26.33   Min.   :91.6    Min.   :1           
#>  1st Qu.:1.468   1st Qu.:  79.20   1st Qu.:91.6    1st Qu.:1           
#>  Median :2.551   Median : 156.93   Median :91.6    Median :1           
#>  Mean   :2.678   Mean   : 297.19   Mean   :91.6    Mean   :1           
#>  3rd Qu.:3.428   3rd Qu.: 307.36   3rd Qu.:91.6    3rd Qu.:1           
#>  Max.   :7.201   Max.   :1979.52   Max.   :91.6    Max.   :1           
#>  ordered_dimension indicator_scale reference_subgroup
#>  Min.   :0         Min.   :100     Min.   :0.00000   
#>  1st Qu.:0         1st Qu.:100     1st Qu.:0.00000   
#>  Median :0         Median :100     Median :0.00000   
#>  Mean   :0         Mean   :100     Mean   :0.02941   
#>  3rd Qu.:0         3rd Qu.:100     3rd Qu.:0.00000   
#>  Max.   :0         Max.   :100     Max.   :1.00000