World Health Organization (WHO)

This dataset contains sample data for computing ordered summary measures of health inequality. It contains data from a household survey for two indicators, the proportion of births attended by skilled health personnel and under-five mortality rate, disaggregated by economic status.

OrderedSampleMultipleind

Format

OrderedSampleMultipleind

A data frame with 10 rows and 11 columns:

indicator

indicator name

dimension

dimension of inequality

subgroup

population subgroup within a given dimension of inequality

subgroup_order

the order of subgroups in an increasing sequence

estimate

subgroup estimate

se

standard error of the subgroup estimate

population

number of people within each subgroup

setting_average

indicator average for the setting

favourable_indicator

favourable (1) or non-favourable (0) indicator

ordered_dimension

ordered (1) or non-ordered (0) dimension

indicator_scale

scale of the indicator

Source

WHO Health Inequality Data Repositoryhttps://www.who.int/data/inequality-monitor/data

Details

The proportion of births attended by skilled health personnel is calculated as the number of births attended by skilled health personnel divided by the total number of live births to women aged 15-49 years occurring in the period prior to the survey.

Skilled health personnel include doctors, nurses, midwives and other medically trained personnel, as defined according to each country. This is in line with the definition used by the Countdown to 2030 Collaboration, Demographic and Health Surveys (DHS), Multiple Indicator Cluster Surveys (MICS) and Reproductive Health Surveys (RHS).

The under-five mortality rate is the probability (expressed as a rate per 1000 live births) of a child born in a specific year or period dying before reaching the age of five years. It is calculated as the number of deaths at age 0-5 years divided by the number of surviving children at the beginning of the specified age range during the 10 years prior to the survey.

Economic status is determined using a wealth index, which is based on owning selected assets and having access to certain services. The wealth index is divided into five equal subgroups (quintiles) that each account for 20% of the population. Economic status is an ordered dimension (meaning that the subgroups have an inherent ordering).

This dataset can be used to calculate ordered summary measures of health inequality, including: absolute concentration index (ACI), relative concentration index (RCI), slope index of inequality (SII) and relative index of inequality (RII). It can also be used to calculate the impact measures population attributable risk (PAR) and population attributable fraction (PAF).

Examples

head(OrderedSampleMultipleind)
#>                                                 indicator
#> 1         Births attended by skilled health personnel (%)
#> 2         Births attended by skilled health personnel (%)
#> 3         Births attended by skilled health personnel (%)
#> 4         Births attended by skilled health personnel (%)
#> 5         Births attended by skilled health personnel (%)
#> 6 Under-five mortality rate (deaths per 1000 live births)
#>                           dimension             subgroup subgroup_order
#> 1 Economic status (wealth quintile) Quintile 1 (poorest)              1
#> 2 Economic status (wealth quintile)           Quintile 2              2
#> 3 Economic status (wealth quintile)           Quintile 3              3
#> 4 Economic status (wealth quintile)           Quintile 4              4
#> 5 Economic status (wealth quintile) Quintile 5 (richest)              5
#> 6 Economic status (wealth quintile) Quintile 1 (poorest)              1
#>   estimate        se population setting_average favourable_indicator
#> 1 75.60530 1.5996131   2072.436        91.59669                    1
#> 2 91.01997 1.1351504   2112.204        91.59669                    1
#> 3 96.03959 0.6461946   1983.059        91.59669                    1
#> 4 97.04223 0.5676206   2052.124        91.59669                    1
#> 5 99.22405 0.2237683   1884.510        91.59669                    1
#> 6 52.46997 3.1404064   7119.732       340.38391                    0
#>   ordered_dimension indicator_scale
#> 1                 1             100
#> 2                 1             100
#> 3                 1             100
#> 4                 1             100
#> 5                 1             100
#> 6                 1            1000
summary(OrderedSampleMultipleind)
#>   indicator          dimension           subgroup         subgroup_order
#>  Length:10          Length:10          Length:10          Min.   :1     
#>  Class :character   Class :character   Class :character   1st Qu.:2     
#>  Mode  :character   Mode  :character   Mode  :character   Median :3     
#>                                                           Mean   :3     
#>                                                           3rd Qu.:4     
#>                                                           Max.   :5     
#>     estimate           se           population   setting_average
#>  Min.   :23.90   Min.   :0.2238   Min.   :1885   Min.   : 91.6  
#>  1st Qu.:31.67   1st Qu.:0.7684   1st Qu.:2057   1st Qu.: 91.6  
#>  Median :64.04   Median :1.9412   Median :4428   Median :216.0  
#>  Mean   :62.84   Mean   :1.7732   Mean   :4472   Mean   :216.0  
#>  3rd Qu.:94.78   3rd Qu.:2.6824   3rd Qu.:6893   3rd Qu.:340.4  
#>  Max.   :99.22   Max.   :3.1404   Max.   :7120   Max.   :340.4  
#>  favourable_indicator ordered_dimension indicator_scale
#>  Min.   :0.0          Min.   :1         Min.   : 100   
#>  1st Qu.:0.0          1st Qu.:1         1st Qu.: 100   
#>  Median :0.5          Median :1         Median : 550   
#>  Mean   :0.5          Mean   :1         Mean   : 550   
#>  3rd Qu.:1.0          3rd Qu.:1         3rd Qu.:1000   
#>  Max.   :1.0          Max.   :1         Max.   :1000