Between-group standard deviation (BGSD)

Between-Group Standard Deviation (BGSD) is an absolute measure of inequality that considers all population subgroups. Subgroups are weighted according to their population share.

bgsd(pop, est, se = NULL, scaleval, sim = NULL, seed = 123456, ...)

Arguments

pop

The number of people within each subgroup. Population size must be available for all subgroups.

est

The subgroup estimate. Estimates must be available for all subgroups.

se

The standard error of the subgroup estimate. If this is missing, 95% confidence intervals of BGSD cannot be calculated.

scaleval

The scale of the indicator. For example, the scale of an indicator measured as a percentage is 100. The scale of an indicator measured as a rate per 1000 population is 1000.

sim

The number of simulations to estimate 95% confidence intervals

seed

The random number generator (RNG) state for the 95% confidence interval simulation

...

Further arguments passed to or from other methods.

Value

The estimated BGSD value, corresponding estimated standard error, and confidence interval as a data.frame.

Details

BGSD is calculated as the square root of the weighted average of squared differences between the subgroup estimates and the setting average. Squared differences are weighted by each subgroup’s population share. For more information on this inequality measure see Schlotheuber, A., & Hosseinpoor, A. R. (2022) below.

95% confidence intervals are calculated using a methodology of simulated estimates. The dataset is simulated a large number of times (e.g., 100) and BGSD is calculated for each of the simulated samples. The 95% confidence intervals are based on the 2.5th and 97.5th percentiles of the BGSD results.

Interpretation: BGSD has only positive values, with larger values indicating higher levels of inequality. BGSD is zero if there is no inequality. It has the same unit as the health indicator.

Type of summary measure: Complex; absolute; weighted

Applicability: Non-ordered; more than two subgroups

References

Schlotheuber, A., & Hosseinpoor, A. R. (2022). Summary measures of health inequality: A review of existing measures and their application. International Journal of Environmental Research and Public Health, 19 (6), 3697.

Examples

# example code
data(NonorderedSample)
head(NonorderedSample)
#> # A tibble: 6 × 11
#>   indicator         dimension subgroup estimate    se population setting_average
#>   <chr>             <chr>     <chr>       <dbl> <dbl>      <dbl>           <dbl>
#> 1 Births attended … Subnatio… aceh         95.1 1.54       230.             91.6
#> 2 Births attended … Subnatio… bali        100   0          149.             91.6
#> 3 Births attended … Subnatio… bangka …     97.4 1.27        55.7            91.6
#> 4 Births attended … Subnatio… banten       80.4 3.54       451.             91.6
#> 5 Births attended … Subnatio… bengkulu     94.3 2.77        70.2            91.6
#> 6 Births attended … Subnatio… central…     98.6 0.648     1222.             91.6
#> # ℹ 4 more variables: favourable_indicator <dbl>, ordered_dimension <dbl>,
#> #   indicator_scale <dbl>, reference_subgroup <dbl>
with(NonorderedSample,
     bgsd(pop = population,
          est = estimate,
          se = se,
          scaleval = indicator_scale
         )
     )
#>      measure estimate  lowerci  upperci
#> 2.5%    bgsd  7.10072 6.327117 8.540935