Difference (D)

The difference (D) is an absolute measure of inequality that shows the difference in a health indicator between two population subgroups. For more information on this inequality measure see Schlotheuber, A., & Hosseinpoor, A. R. (2022) below.

Usage

d(
  est,
  se,
  favourable_indicator,
  ordered_dimension = NULL,
  subgroup_order = NULL,
  reference_subgroup = NULL,
  conf.level = 0.95,
  ...
)

Arguments

est

The subgroup estimate. Estimates must be available for all subgroups.

se

The standard error of the subgroup estimate. If this is missing, 95% confidence intervals of MDBU cannot be calculated.

favourable_indicator

Records whether the indicator is favourable (1) or non-favourable (0). Favourable indicators measure desirable health events where the ultimate goal is to achieve a maximum level (such as skilled birth attendance). Non-favourable indicators measure undesirable health events where the ultimate goal is to achieve a minimum level (such as under-five mortality rate).

ordered_dimension

Records whether the dimension is ordered (1) or not (0).

subgroup_order

The order of subgroups in an increasing sequence.

reference_subgroup

Identifies a reference subgroup with the value of 1.

conf.level

confidence level of the interval.

...

Further arguments passed to or from other methods.

Value

The estimated D value, corresponding estimated standard error, and confidence interval as a data.frame.

Details

D is calculated as: D = y_1 - y_2, where y_1 and y_2 indicate the estimates for subgroups 1 and 2. The selection of the two subgroups depends on the characteristics of the inequality dimension and the purpose of the analysis. In addition, the direction of the calculation may depend on the indicator type (favourable or#' adverse).

Interpretation: Greater absolute values indicate higher levels of inequality. D is zero if there is no inequality.

Type of summary measure: Simple; relative; unweighted

Applicability: Any

Warning: The confidence intervals are approximate and might be biased.

References

Schlotheuber, A., & Hosseinpoor, A. R. (2022). Summary measures of health inequality: A review of existing measures and their application. International Journal of Environmental Research and Public Health, 19 (6), 3697.

Examples

# example code
data(NonorderedSample)
head(NonorderedSample)
#> # A tibble: 6 × 11
#>   indicator         dimension subgroup estimate    se population setting_average
#>   <chr>             <chr>     <chr>       <dbl> <dbl>      <dbl>           <dbl>
#> 1 Births attended … Subnatio… aceh         95.1 1.54       230.             91.6
#> 2 Births attended … Subnatio… bali        100   0          149.             91.6
#> 3 Births attended … Subnatio… bangka …     97.4 1.27        55.7            91.6
#> 4 Births attended … Subnatio… banten       80.4 3.54       451.             91.6
#> 5 Births attended … Subnatio… bengkulu     94.3 2.77        70.2            91.6
#> 6 Births attended … Subnatio… central…     98.6 0.648     1222.             91.6
#> # ℹ 4 more variables: favourable_indicator <dbl>, ordered_dimension <dbl>,
#> #   indicator_scale <dbl>, reference_subgroup <dbl>
with(NonorderedSample,
        d(est = estimate,
          se = se,
          favourable_indicator = favourable_indicator,
          ordered_dimension = ordered_dimension,
          reference_subgroup = reference_subgroup
         )
     )
#>   measure estimate       se lowerci  upperci
#> 1       d 30.92258 7.217724 16.7761 45.06906