Relative concentration index (RCI)

The relative concentration index (RCI) is a relative measure of inequality that shows the gradient across population subgroups. It indicates the extent to which an indicator is concentrated among disadvantaged or advantaged subgroups, on a relative scale. RCI can be calculated using disaggregated data and individual-level data. Subgroups in disaggregated data are weighted according to their population share, while individuals are weighted by sample weight in the case of data from surveys.

Usage

rci(
  est,
  subgroup_order,
  scaleval = NULL,
  pop = NULL,
  weight = NULL,
  psu = NULL,
  strata = NULL,
  fpc = NULL,
  method = NULL,
  lmin = NULL,
  lmax = NULL,
  conf.level = 0.95,
  force = FALSE,
  ...
)

Arguments

est

The indicator estimate. Estimates must be available for all subgroups/individuals (unless force=TRUE).

subgroup_order

The order of subgroups/individuals in an increasing sequence.

scaleval

The scale of the indicator. For example, the scale of an indicator measured as a percentage is 100. The scale of an indicator measured as a rate per 1000 population is 1000.

pop

The number of people within each subgroup (for disaggregated data). Population size must be available for all subgroups.

weight

Individual sampling weight (required if data come from a survey)

psu

Primary sampling unit (required if data come from a survey)

strata

Strata (required if data come from a survey)

fpc

Finite population correction (if data come from a survey and sample size is large relative to population size).

method

Normalisation method for bounded indicators. Options available Wagstaff (wagstaff) or Erreygers (erreygers) correction.

lmin

Minimum limit for bounded indicators (i.e., variables that have a finite upper and/or lower limit).

lmax

Maximum limit for bounded indicators (i.e., variables that have a finite upper and/or lower limit).

conf.level

Confidence level of the interval. Default is 0.95 (95%).

force

TRUE/FALSE statement to force calculation with missing indicator estimate values.

...

Further arguments passed to or from other methods.

Value

The estimated RCI value, corresponding estimated standard error, and confidence interval as a data.frame.

Details

The calculation of RCI is based on a ranking of the whole population from the most-disadvantaged subgroup (at rank 0) to the most-advantaged subgroup (at rank 1), which is inferred from the ranking and size of the subgroups. RCI can be calculated as twice the covariance between the health indicator and the relative rank, divided by the indicator mean. Given the relationship between covariance and ordinary least squares regression, RCI can be obtained from a regression of a transformation of the health variable of interest on the relative rank. For more information on this inequality measure see Schlotheuber, A., & Hosseinpoor, A. R. (2022) below.

Interpretation: RCI is bounded between -1 and +1 (or between -100 and +100, when multiplied by 100). The larger the absolute value of RCI, the higher the level of inequality. For favourable indicators, positive values indicate a concentration of the indicator among the advantaged, while negative values indicate a concentration of the indicator among the disadvantaged. For adverse indicators, it is the reverse: positive values indicate a concentration of the indicator among the disadvantaged, while negative values indicate a concentration of the indicator among the advantaged. RCI is zero if there is no inequality.

Type of summary measure: Complex; relative; weighted

Applicability: Ordered; more than two subgroups

Warning: The confidence intervals are approximate and might be biased.

References

Schlotheuber, A., & Hosseinpoor, A. R. (2022). Summary measures of health inequality: A review of existing measures and their application. International Journal of Environmental Research and Public Health, 19 (6), 3697.

Wagstaff A. (2005). The bounds of the concentration index when the variable of interest is binary, with an application to immunization inequality. Health Economics, 14:429–432.

Erreygers G. (2009). Correcting the Concentration Index. Journal of Health Economics, 28:504–515.

Examples

# example code
data(IndividualSample)
head(IndividualSample)
#>   id  psu strata   weight             subgroup subgroup_order sba dtp3
#> 1  1   88      1 0.351672           Quintile 3              3   1   NA
#> 2  2 1337     38 0.431545 Quintile 1 (poorest)              1   1    1
#> 3  3  450     18 0.482483           Quintile 2              2   1   NA
#> 4  4 1692     56 0.407390 Quintile 1 (poorest)              1   0   NA
#> 5  5  752     23 1.547062 Quintile 5 (richest)              5   1   NA
#> 6  6 1033     30 2.429523 Quintile 5 (richest)              5   1   NA
#>   favourable_indicator indicator_scale
#> 1                    1             100
#> 2                    1             100
#> 3                    1             100
#> 4                    1             100
#> 5                    1             100
#> 6                    1             100
with(IndividualSample,
     rci(est = sba,
         subgroup_order = subgroup_order,
         weight = weight,
         psu = psu,
         strata = strata
         )
     )
#>   measure   estimate          se    lowerci  upperci
#> 1     rci 0.09723792 0.003677656 0.09002985 0.104446